+++ src/commands/compileScript.ts
@@ -11,6 +11,10 @@ export default class CompileScript {
       new CompileScript().transpile(document, m[1] !== undefined);
     }
   }
+  static executeRename(e: vscode.FileRenameEvent) {
+    let files = e.files.filter(f => f.oldUri.fsPath.search(/\.script(library)?\.ts$/) > -1);
+    files.forEach(f => new CompileScript().rename(f.oldUri, f.newUri));
+  }
 
   private async transpile(document: vscode.TextDocument, isLib: boolean) {
     let content = document.getText();
@@ -27,6 +31,16 @@ export default class CompileScript {
     let jsContent = ts.transpileModule(content, { compilerOptions: { target: ts.ScriptTarget.ES5 } });
     let bin = new TextEncoder().encode(jsContent.outputText);
     await vscode.workspace.fs.writeFile(outFile, bin);
+  }
+
+  private async rename(oldUri: vscode.Uri, newUri: vscode.Uri) {
+    let oldJs = vscode.Uri.file(this.getCompiledJsFileName(oldUri.fsPath));
+    let newJs = vscode.Uri.file(this.getCompiledJsFileName(newUri.fsPath));
+    try {
+      await vscode.workspace.fs.rename(oldJs, newJs);
+    }
+    catch {
+    }
   }
 
   private getCompiledJsFileName(tsFileName: string) {